게시글 삭제 기능 구현, 조회수 증가 로직 변경

+++ src/main/java/yougboyclub/honbabstop/service/BoardServiceImpl.java
@@ -101,11 +101,11 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public Board findByIdAndUser(Long id, User currentUser) {
         Board board = boardRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("찾지 못했습니다: " + id));
-
         // 본인 게시글이 아닐 경우에만 조회수 증가
         if (!board.getWriter().getId().equals(currentUser.getId())) {
             board.increaseHit();
         }
+        boardRepository.save(board);
         return board;
     }
 
@@ -149,8 +149,13 @@ public class BoardServiceImpl implements BoardService {
   //특정 모집글 삭제
   @Override
   @Transactional
-  public void deleteById(Long id) {
+  public void deleteById(Long id, Long userId) {
+    Board board = boardRepository.findById(id)
+        .orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id=" + id));
+    // 권한 확인
+    if (!board.getWriter().getId().equals(userId)) {
+      throw new IllegalArgumentException("게시글 삭제 권한이 없습니다.");
+    }
     boardRepository.deleteById(id);
   }
-
 }
